Refactor booking form into subcomponents
